Move injection of transaction up to ContentExposingResource so that it does not automatically happen in Transactions controller. Moved methods which depend on transaction with it.
